third_party/WebKit/LayoutTests/animations/transition-and-animation-3.html

<!DOCTYPE html>

<html>
<head>
    <style type="text/css" media="screen">
        .box {
            position: absolute;
            height: 100px;
            width: 100px;
            left: 0px;
            background-color: blue;
        }

        .box.animating {
            animation: move 0.1s linear;
            left: 100px;
            transition: left 10s linear;
        }

        .box.animating.moved {
            left: 200px;
        }
        
        @keyframes move {
            from {
                left: 500px;
            }
            to   {
                left: 501px;
            }
        }
        
        #result {
            margin-top: 130px;
        }
    </style>
    <script src="resources/animation-test-helpers.js" type="text/javascript" charset="utf-8"></script>
    <script type="text/javascript" charset="utf-8">

    if (window.testRunner)
        testRunner.waitUntilDone();

      const expectedValues = [
        // [time, element-id, property, expected-value, tolerance]
        [0.2, "test", "left", 200, 10],
      ];

      function animationStarted()
      {
          document.getElementById('test').className = 'animating moved box';
      }

      function setupTest()
      {
        document.body.offsetHeight; // Force style recalc
        document.getElementById('test').className = 'animating box';
        runAnimationTest(expectedValues, animationStarted, undefined, 'do-not-use-pause-api');
      }
      
      window.addEventListener('load', setupTest, false);

    </script>
</head>
<body>
    <p>Once animation has finished, box shouldn't transitioning.</p>
    <div id="test" class="box"></div>
    <div id="result"></div>
</body>
</html>
